requirements.txt
# Web3 / blockchain
web3>=6.0.0
eth-account>=0.10.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
